requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
google-cloud-firestore>=2.16.0
scikit-learn>=1.3.0
joblib>=1.3.0
pandas>=2.0.0
xgboost>=2.0.0
chainlit>=1.0.0
